Fix Implicit Conversion Warning in SFTPBrowserDlg::ClearView
Add an explicit `static_cast<size_t>` to the return value of `m_dataview->GetItemCount()` in the loop condition to resolve a signed/unsigned comparison warning.

* Plugin/SFTPBrowserDlg.cpp

+++ Plugin/SFTPBrowserDlg.cpp
@@ -404,7 +404,7 @@ void SFTPBrowserDlg::OnSSHAccountManager(wxCommandEvent& event)
 
 void SFTPBrowserDlg::ClearView()
 {
-    for (size_t i = 0; i < m_dataview->GetItemCount(); ++i) {
+    for (size_t i = 0; i < static_cast<size_t>(m_dataview->GetItemCount()); ++i) {
         SFTPBrowserEntryClientData* cd = DoGetItemData(m_dataview->RowToItem(i));
         wxDELETE(cd);
     }
